add variadic logging

+++ src/wrap.h
@@ -8,7 +8,7 @@
 using std::string;
 
 void clearBackground(colorRGB col);
-void drawRectangle(int x, int y, int w, int h, colorRGB col);
+void drawRectangle(float x, float y, float w, float h, colorRGB col);
 void drawLine(int xi, int yi, int xf, int yf, colorRGB col);
 void drawLineEx(int xi, int yi, int xf, int yf, float thick, colorRGB col);
 void drawTextEx(Font ft, string msg, int x, int y, colorRGB col);
